2018-9-20 Fix some bugs.

+++ Abp.EntityFramework/Entities/RegionKpi/TownPreciseView.cs
@@ -21,11 +21,11 @@ namespace Abp.EntityFramework.Entities.RegionKpi
 
         public FrequencyBandType FrequencyBandType { get; set; }
 
-        public int TotalMrs { get; set; }
+        public long TotalMrs { get; set; }
 
-        public int ThirdNeighbors { get; set; }
+        public long ThirdNeighbors { get; set; }
 
-        public int SecondNeighbors { get; set; }
+        public long SecondNeighbors { get; set; }
 
         public int FirstNeighbors { get; set; }
 
@@ -35,12 +35,12 @@ namespace Abp.EntityFramework.Entities.RegionKpi
 
         public double ThirdRate => 100 - (double)ThirdNeighbors * 100 / TotalMrs;
 
-        public int NeighborsMore { get; set; }
+        public long NeighborsMore { get; set; }
 
-        public int InterFirstNeighbors { get; set; }
+        public long InterFirstNeighbors { get; set; }
 
-        public int InterSecondNeighbors { get; set; }
+        public long InterSecondNeighbors { get; set; }
 
-        public int InterThirdNeighbors { get; set; }
+        public long InterThirdNeighbors { get; set; }
     }
 }
